requirements.txt
discord.py
aiohttp
flask
beautifulsoup4
lxml
psutil
dotenv